+++ Source/ACE.Database/CharacterDatabase.cs
@@ -279,26 +279,7 @@ namespace ACE.Database
                 LoadCharacterPositions(c);
 
                 // Load the best location from the database
-                // This may be needed if a position is removed from the database, after a character has been created.
-                // But should now be totally un-neccssary.
-                var invalidPosition = CharacterPositionExtensions.InvalidPosition(c.Id, PositionType.Undef);
-
-                if (c.Positions.ContainsKey(PositionType.Location))
-                    c.Location = c.Positions[PositionType.Location];
-                else if (c.Positions.ContainsKey(PositionType.LastPortal))
-                    c.Location = c.Positions[PositionType.LastPortal];
-                else if (c.Positions.ContainsKey(PositionType.Sanctuary))
-                    c.Location = c.Positions[PositionType.Sanctuary];
-
-                // Check to see if the database position is valid
-                if (c.Location.PositionX == invalidPosition.PositionX && 
-                    c.Location.PositionY == invalidPosition.PositionY && 
-                    c.Location.PositionZ == invalidPosition.PositionZ)
-                {
-                    // If the cell is also invalid, we will set to initial character starting location.
-                    if (c.Location.LandblockId.Raw == 0x0 || c.Location.LandblockId.Raw == 0x255)
-                        c.Location = CharacterPositionExtensions.StartingPosition(c.Id);
-                }
+                c.Location = c.Positions[PositionType.Location];
 
                 uint characterOptions1Flag = result.Read<uint>(0, "characterOptions1");
                 uint characterOptions2Flag = result.Read<uint>(0, "characterOptions2");
@@ -446,13 +427,27 @@ namespace ACE.Database
                 dbObject.SetPropertyString(results.Read<PropertyString>(i, "propertyId"), results.Read<string>(i, "propertyValue"));
         }
 
-        public List<Position> GetCharacterPositions(Character character)
+        /// <summary>
+        /// Returns all positions from the database in a dictionary. Positions for each character should always be unique.
+        /// </summary>
+        public Dictionary<PositionType, Position> GetCharacterPositions(Character character)
         {
             Dictionary<string, object> criteria = new Dictionary<string, object>();
             criteria.Add("character_id", character.Id);
-            return ExecuteConstructedGetListStatement<CharacterPreparedStatement, Position>(CharacterPreparedStatement.CharacterPositionList, criteria);
+            List<Position> availablePositions = ExecuteConstructedGetListStatement<CharacterPreparedStatement, Position>(CharacterPreparedStatement.CharacterPositionList, criteria);
+            Dictionary<PositionType, Position> characterPositions = new Dictionary<PositionType, Position>();
+            foreach (Position position in availablePositions)
+            {
+                characterPositions.Add(position.PositionType, position);
+            }
+
+            return characterPositions;
         }
 
+        /// <summary>
+        /// Returns an available character positions from the database that match the PositionType.
+        /// If no position is found, an invalid position is returned.
+        /// </summary>
         public Position GetCharacterPosition(Character character, PositionType positionType)
         {
             Dictionary<string, object> criteria = new Dictionary<string, object>();
@@ -467,35 +462,44 @@ namespace ACE.Database
         /// <summary>
         /// Attempts to load characters positions from the database. If no position is available, a new position will be created to allow the player to spawn.
         /// </summary>
+        /// <remarks>
+        /// TODO:
+        ///     1. Confirm the order/position usage when restoring a position from database. Should this always be the lifestone if "Location" is unavailable?
+        /// </remarks>
         public void LoadCharacterPositions(Character character)
         {
             // get a list of positions from the vw_character_positions view
-            var dbPositionList = GetCharacterPositions(character);
+            var dbPositions = GetCharacterPositions(character);
 
-            bool containsLocation = false;
-
-            if (dbPositionList.Count > 0)
+            // Check for positions and insert position for the Position type Locaiont, if missing:
+            if (dbPositions.Count == 0 || !dbPositions.ContainsKey(PositionType.Location))
             {
-                foreach (Position checkPosition in dbPositionList)
+                Position swapToWorkingLocation = new Position();
+
+                // This checks for the best location available from the database
+                // When the first suitable position is found, it will be set as the current location.
+                if (dbPositions.ContainsKey(PositionType.LastPortal)) 
+                    // Use the lastportal position
+                    swapToWorkingLocation = dbPositions[PositionType.LastPortal]; 
+                else if (dbPositions.ContainsKey(PositionType.Sanctuary)) 
+                    // Use lifestone recall position
+                    swapToWorkingLocation = dbPositions[PositionType.Sanctuary]; 
+                else
                 {
-                    if (checkPosition.PositionType == PositionType.Location)
-                        containsLocation = true;
+                    // No suitable database position is available
+                    swapToWorkingLocation = CharacterPositionExtensions.StartingPosition(character.Id); //Use default position
                 }
-            }
 
-            // Check for positions and insert defaults if missing:
-            if (dbPositionList.Count == 0 || !containsLocation)
-            {
-                // load the default position
-                Position newCharacterPosition = CharacterPositionExtensions.StartingPosition(character.Id);
-                dbPositionList.Add(newCharacterPosition);
-                ExecuteConstructedInsertStatement(CharacterPreparedStatement.CharacterPositionInsert, typeof(Position), newCharacterPosition);
+                // force the position type to location
+                swapToWorkingLocation.PositionType = PositionType.Location;
+                dbPositions.Add(swapToWorkingLocation.PositionType, swapToWorkingLocation);
+                ExecuteConstructedInsertStatement(CharacterPreparedStatement.CharacterPositionInsert, typeof(Position), swapToWorkingLocation);
             }
 
             // This will load each available position from the database, into the object's positions
-            foreach (Position item in dbPositionList)
+            foreach (KeyValuePair<PositionType, Position> position in dbPositions)
             {
-                character.SetCharacterPosition(item);
+                character.SetCharacterPosition(position.Value);
             }
         }